Remove mysql alias, import .my.cnf dotfile
Aliasing mysql for having some default arguments applied was a bad idea.
In particular it breaks arguements like `--print-defaults`, `--no-defaults` and
`--login-path` which have to be given as the very first argument.

But configs can be also applied vie .my.cnf and this is a perfect candidate for
a new dotfile.