Remove DB host envs
Remove MYSQL_HOST env which was explicitly set to 127.0.0.1 It was set to
127.0.0.1 so the client was forced to connect via TCP and not via a socket
connection, because I was used to do local port forwarding via SSH and just had
to remember which port was pointing to which host.

However, this became obsolete with `mysql_config_editor` (which I wasn't aware
of by that time). At mycli, an env takes precedence over `--login-path` configs.