composer: make sure, config and cache dir a created by user
In case the command is executed for the first time and `~/.config/composer` or
`~/.cache/composer` are not present, which causes the docker daemon to create
these on demand. However, doing so, will make them belong to the root user.

As such, we make sure the directories are always present and writeable by the
executing user.